Milestone-5/backend/requirements.txt
alembic==1.10.2
amqp==5.1.1
aniso8601==9.0.1
async-timeout==4.0.2
billiard==3.6.4.0
cachetools==5.3.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
cryptography==40.0.1
Flask==2.2.3
Flask-Cors==3.0.10
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
oauthlib==3.2.2
prompt-toolkit==3.0.38
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytz==2023.3
redis==4.5.4
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
pytest==7.3.1